requirements.txt
streamlit>=1.36
Pillow>=10
polars>=1.5
pyarrow>=17
pandas>=2.2
openpyxl>=3.1
plotly>=5.22
python-dateutil>=2.9
